+++ contracts/helpers/IERC1155TokenReceiver.sol
@@ -16,7 +16,7 @@ interface IERC1155TokenReceiver {
         @param _data      Additional data with no specified format
         @return           `bytes4(keccak256("onERC1155Received(address,address,uint256,uint256,bytes)"))`
     */
-    function onERC1155Received(address _operator, address _from, uint256 _id, uint256 _value, bytes calldata _data) external returns(bytes4);
+    function onERC1155Received(address _operator, address _from, uint _id, uint _value, bytes calldata _data) external returns(bytes4);
 
     /*
         @notice Handle the receipt of multiple ERC1155 token types.
@@ -31,6 +31,6 @@ interface IERC1155TokenReceiver {
         @param _data      Additional data with no specified format
         @return           `bytes4(keccak256("onERC1155BatchReceived(address,address,uint256[],uint256[],bytes)"))`
     */
-    function onERC1155BatchReceived(address _operator, address _from, uint256[] calldata _ids, uint256[] calldata _values, bytes calldata _data) external returns(bytes4);
+    function onERC1155BatchReceived(address _operator, address _from, uint[] calldata _ids, uint[] calldata _values, bytes calldata _data) external returns(bytes4);
 
 }
