+++ contracts/interfaces/IBank.sol
@@ -8,6 +8,6 @@ interface IBank {
     */
     function tokenAddress() external view returns(address);
 
-    function withdraw(address recipient, uint256 amount) external;
+    function withdraw(address recipient, uint amount) external;
 
 }
